Checkpoint #3-4: More foolproof file removal
  Vim:
  - Improve toggle for numbers and signs to also enable/disable diagnostic.
    Now it works for only the current buffer.
  - Add forgotten alias in the nvim-tree file.
  - Fix path mess with copying nvim's lua files.
  - Fix break-line mapping.
  - Set up (undo/backup)dirs.

  Shell:
  - Give a user time before deleting files with `-rf` flag.
    Ask them when deleting recursively or more than 3 files (if w/o -f option).
  - Add XDG_STATE_HOME variable.
  - Generalize anacron setup script and the job-script it runs.

  - Update README

+++ conf/bash/aliases-functions.sh
@@ -255,8 +255,33 @@ swap () {
 }
 
 
+bak () {
+  for file in "$@"; do
+    [[ -e bak.$file ]] && { echo "bak.$file already exists"; return 1; }
+    cp -r "$file" "bak.$file"
+  done
+}
+
+
+_think_before () {
+  local nsec=$1
+  while [[ $(( nsec -= 1 )) -gt 0 ]]; do
+    echo "You have $nsec second(s) to change your mind"
+    sleep 1
+  done
+}
+
+
+rm () {
+  if [[ " $* " =~ ' -rf ' ]]; then
+    _think_before 5
+  fi
+  command rm -I "$@"
+}
+
+
 vrmswp () {
-  [[ -z $1 ]] && "Pass the name of swap file to delete."
+  [[ -z $1 ]] && { echo "Pass the name of swap file to delete."; return 1; }
   local swp=${1//\//%}
   rm "$XDG_DATA_HOME/nvim/swap/"*$swp*
 }
@@ -277,5 +302,5 @@ join_by () {
   && grep -q '^source .*ipython\.vim' "$XDG_CONFIG_HOME/nvim/init.vim") \
   &> /dev/null && source "$EVANGELIST/conf/tmux/templates.sh"
 
-## Save to use from the interective shell?
+## Save to use from the interactive shell?
 # unset _shell
